requirements.txt
flask-socketio
flask-cors
sqlalchemy
flask-sqlalchemy
flask_login
pandas
matplotlib
passlib
